perf(distill): stop a skeleton import dragging networkx and structlog along
`import repowise.core.distill.skeleton` cost 556ms. None of it was the
skeleton: a package __init__ runs on any submodule import, so the engine's
graph was charged to every importer, and `budget.py` reached into
`generation.context` for a four-line arithmetic heuristic, which runs that
package's assembler -> ingestion models -> networkx.

The PostToolUse Read hook pays this on every fire that reaches the skeleton
path, and it is about to do more there, not less.

Two changes, both structural rather than clever:

- The engine moves to distill/engine.py verbatim. `__init__` becomes a lazy
  PEP 562 re-export shim importing nothing at module scope, so
  `from repowise.core.distill import distill_output` costs exactly what it
  did while importing a sibling submodule costs only that sibling.
- `estimate_tokens` now lives in distill/budget.py and
  generation.context.token_budget re-exports it. Ownership follows the
  import graph: distill is a leaf, generation.context is not.

556ms -> 27ms.
